cambios en Controllador cancion para checkear session

+++ Controllers/CancionController.php
@@ -3,29 +3,34 @@ require_once 'Controller.php';
 require_once '.\Models\CancionModel.php';
 require_once '.\Views\CancionView.php';
 
-class CancionController extends Controller{
+class CancionController extends Controller
+{
 
-    function __construct() {
+    function __construct()
+    {
         $this->model = new CancionModel();
         $this->view = new CancionView();
     }
 
-    public function create() {
-        if(isset($_POST) && isset($_POST['nombre']) && isset($_POST['duracion']) && isset($_POST['genero']) && isset($_POST['album']) && isset($_POST['artista']) && isset($_POST['ranking'])) {
+    public function create()
+    {
+        if (isset($_POST) && isset($_POST['nombre']) && isset($_POST['duracion']) && isset($_POST['genero']) && isset($_POST['album']) && isset($_POST['artista']) && isset($_POST['ranking'])) {
             $this->model->create(array($_POST['nombre'], $_POST['duracion'], $_POST['genero'], $_POST['album'], $_POST['artista'], $_POST['ranking']));
         }
         header("Location: " . BASE_CANCION);
     }
 
-    public function deleter() {
-        if(isset($_POST) && isset($_POST['id'])) {
+    public function deleter()
+    {
+        if (isset($_POST) && isset($_POST['id'])) {
             $this->model->delete($_POST['id']);
         }
         header("Location: " . BASE_CANCION);
     }
 
-    public function update() {
-        if(isset($_POST) && isset($_POST['id']) && isset($_POST['nombre']) && isset($_POST['duracion']) && isset($_POST['genero']) && isset($_POST['album']) && isset($_POST['artista']) && isset($_POST['ranking'])) {
+    public function update()
+    {
+        if (isset($_POST) && isset($_POST['id']) && isset($_POST['nombre']) && isset($_POST['duracion']) && isset($_POST['genero']) && isset($_POST['album']) && isset($_POST['artista']) && isset($_POST['ranking'])) {
             $this->model->update(array($_POST['nombre'], $_POST['duracion'], $_POST['genero'], $_POST['album'], $_POST['artista'], $_POST['ranking'], $_POST['id']));
             header("Location: " . BASE_CANCION);
         } elseif (isset($_GET) && isset($_GET['id'])) {
@@ -34,13 +39,16 @@ class CancionController extends Controller{
         }
     }
 
-    private function findById($id) {
+    private function findById($id)
+    {
         return $this->model->getByID($id);
     }
 
-    public function findByColumn($column,$parameter) {
-        $obj = $this->model->findByColumn($column,$parameter);
-        var_dump($obj);die;
+    public function findByColumn($column, $parameter)
+    {
+        $obj = $this->model->findByColumn($column, $parameter);
+        var_dump($obj);
+        die;
     }
 
 
@@ -53,33 +61,48 @@ class CancionController extends Controller{
 
     public function show($id)
     {
-        $cancion = $this->model->findById($id[]);
+
+        $cancion = $this->model->findById($id);
 
         $this->view->showOne($cancion);
     }
     public function edit($id)
 
     {
-
-        $cancion = $this->model->getByID($id);
-
-        $this->view->edit($cancion);
+        if ($this->view->returnSession()->isLoggedIn()) {
+            $cancion = $this->model->getByID($id);
+            $this->view->edit($cancion);
+        } else {
+            Route::directDefault();
+        }
     }
     public function save($data)
     {
-        $this->model->create($data);
-        $this->index();
+        if ($this->view->returnSession()->isLoggedIn()) {
+            $this->model->create($data);
+            $this->index();
+        } else {
+            Route::directDefault();
+        }
     }
 
     public function insert($data)
     {
-        $this->model->insert($data);
-        $this->index();
+        if ($this->view->returnSession()->isLoggedIn()) {
+            $this->model->insert($data);
+            $this->index();
+        } else {
+            Route::directDefault();
+        }
     }
 
     public function delete($id)
     {
-        $this->model->delete($id);
-        $this->index();
+        if ($this->view->returnSession()->isLoggedIn()) {
+            $this->model->delete($id);
+            $this->index();
+        } else {
+            Route::directDefault();
+        }
     }
 }
